Lot of little bug fixes/useability improvements - the ProgressOverlayViewController seems to work properly now! (I needed to call it from the viewDidAppear method not the viewWillAppear method)...
The Bills area now has the ability to drill-down on a bill to get more info. The pretty GUI formatting isn't there yet, but the data is. This area also functions similarly to the Congress area in that it conforms to the UI guidelines (row-tap brings up action sheet, drill-down view is black, grouped table, etc).

The MiniBrowser seems to occasionally crash - still don't know why, but it doesn't appear to be anything I'm doing and/or handling...