Addressed the following list generated by Christian Bell: [Community] Tapping on bill should offer a contextual 'back' button to return to the community item (this happens if the Community tab is clicked, but this isn't obvious; the Bills back button should be changed to 'Community' instead, and return you to the viewed item). Real names should be displayed in place of usernames. In the simulator, tapping outside the keyboard when posting a reply does not dismiss the keyboard. Not sure if this happens on-device. Signing in with Google credentials posts comments as 'Anonymous' initially. I posted a comment under Pete Hoekstra and it showed as anonymous. It now correctly shows as cshbell. The Chatter button does nothing when pressed; this should not be a button until it does something.
[Recovery]
$10 says the AppStore reviewers will not like the Recovery page. For now, it should display the recovery.gov webpage with an overlay stating that we're waiting for feeds to become available in summer 2009.

[Congress]
Twitter DM button should not appear unless the candidate has a Twitter account.
The 'Twitter DM Error' pop-up is ambiguous (that is, the error you receive when attempting to DM a legislator who doesn't follow you). The two buttons are OK and Reset. Which one does what? They seem to do the same thing. Confusing.
More to the point: Why is the first option to DM a legislator rather than simply @ tweeting them? Replies are much more common than DMs, and won't run into the "are they or aren't they following you?" problem. I'd suggest scrapping DMing and making the DM feature @ replies instead. (Especially since Twitter just overhauled their site to change 'replies' to 'mentions')
When searching for a congressperson, tapping outside the keyboard does not dismiss the keyboard (it is only dismissed if the Cancel button next to the search field is tapped). Again, could be a simulator bug.

[Bills]
As above in the first item under [Community], tapping a bill sponsor takes you to the congressperson's myGov page, but there is not a contextual 'Back' button, but instead a back button to go to the Congress section. I again feel that this is confusing.
Minor UI quibble: The green for 'Passed' legislation is a bit bright; looks like #00ff00; maybe something more like #00b100?