fixed a bills-tab related crash
Occasionally the bills tab would come up with 0 bills, but would still try to index to 0,0 which didn't exist - I tried to guard against this.

There is still a funny issue on the simulator when I try to re-download bill data. If I select to delete existing data and re-download everything, but don't have anything currently in my bill cache, things seem to stall... haven't traced that one through yet...