Bill data cache load is much faster now (esp. on older devices)
I broke up the bill data cache into months for both house and sentate, then allowed the TableViewController to load whatever data is currently available. This allows the user to start interacting with bill data much faster (they don't have to wait for _all_ of it to load). It loads from youngest to oldest. 

Also added a couple of UI features for bill reloading progress indication, and fixed a couple of nasty crashing bugs when a bioguideID on a bill sponsor didn't match anything in the congress database... oops :-)